+++ modules/agar/src/main/ts/ephox/agar/find/UiSearcher.ts
@@ -1,8 +1,7 @@
 import { Adt, Option, Result } from '@ephox/katamari';
-import { Element } from '@ephox/sugar';
+import { Element, Truncate } from '@ephox/sugar';
 
 import * as SizzleFind from '../alien/SizzleFind';
-import * as Truncate from '../alien/Truncate';
 import { TestLabel } from '@ephox/bedrock-client';
 
 interface TargetAdt {